feat: implemented AverageSpeed.kt to compute average speed of the route   feat: implemented CardinalDirection.kt to compute the direction of the movement   refactor: haversine now accepts waypoints instead of coordinates   fix: fixed Parameters.kt wanting Int instead of Double as timestamp

+++ RouteAnalyzer/src/main/kotlin/models/Waypoint.kt
@@ -6,4 +6,4 @@ package org.example.models
  * @property latitude the latitude of the waypoint
  * @property longitude the longitude of the waypoint
  */
-data class Waypoint(val timestamp: Int, val latitude: Double, val longitude: Double)
+data class Waypoint(val timestamp: Double, val latitude: Double, val longitude: Double)
